requirements/testing.txt
-r base.txt
pytest-django==3.9.0
pytest==5.4.3
pytest-cov==2.9.0
